actually add initial tests

+++ src/trackinghash.h
@@ -16,7 +16,7 @@
     size_t idx_max;            // how big the contents are
   };
   struct track_hash * VALC_create_track_hash(size_t size_init);
-  int VALC_add_to_track_hash(
+  size_t VALC_add_to_track_hash(
     struct track_hash * track_hash, const char * key, const char * value
   );
   void VALC_reset_track_hash(struct track_hash * track_hash, size_t idx);
